Server tests now Typescript files and moved then alongside the original file.
