--- server/tests/sample.test.js
@@ -1,5 +1,0 @@
-describe("Sample Test", () => {
-    it("should test that true === true", () => {
-        expect(true).toBe(true);
-    });
-});
